Refactored app into split files
