+++ main.go
@@ -2,208 +2,21 @@ package main
 
 import (
 	"bytes"
-	"compress/gzip"
-	"embed"
 	"encoding/json"
 	"fmt"
-	"io"
 	"io/fs"
 	"os"
 	"path/filepath"
 	"runtime"
 	"slices"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
-	"time"
 
 	check "github.com/andreimerlescu/checkfs"
 	"github.com/andreimerlescu/checkfs/directory"
-	"github.com/andreimerlescu/checkfs/file"
-	"github.com/andreimerlescu/figtree/v2"
 	"github.com/andreimerlescu/sema"
 )
-
-//go:embed VERSION
-var versionBytes embed.FS
-
-var currentVersion string
-
-func Version() string {
-	if len(currentVersion) == 0 {
-		versionBytes, err := versionBytes.ReadFile("VERSION")
-		if err != nil {
-			return ""
-		}
-		currentVersion = strings.TrimSpace(string(versionBytes))
-	}
-	return currentVersion
-}
-
-const (
-	projectName string = "github.com/andreimerlescu/summarize"
-	tFormat     string = "2006.01.02.15.04.05.UTC"
-
-	eConfigFile          string = "SUMMARIZE_CONFIG_FILE"
-	eAddIgnoreInPathList string = "SUMMARIZE_IGNORE_CONTAINS"
-	eAddIncludeExtList   string = "SUMMARIZE_INCLUDE_EXT"
-	eAddExcludeExtList   string = "SUMMARIZE_EXCLUDE_EXT"
-	eAlwaysWrite         string = "SUMMARIZE_ALWAYS_WRITE"
-	eAlwaysPrint         string = "SUMMARIZE_ALWAYS_PRINT"
-	eAlwaysJson          string = "SUMMARIZE_ALWAYS_JSON"
-	eAlwaysCompress      string = "SUMMARIZE_ALWAYS_COMPRESS"
-
-	kSourceDir     string = "d"
-	kOutputDir     string = "o"
-	kIncludeExt    string = "i"
-	kExcludeExt    string = "x"
-	kSkipContains  string = "s"
-	kFilename      string = "f"
-	kPrint         string = "print"
-	kMaxOutputSize string = "max"
-	kWrite         string = "write"
-	kVersion       string = "v"
-	kDotFiles      string = "ndf"
-	kMaxFiles      string = "mf"
-	kDebug         string = "debug"
-	kJson          string = "json"
-	kCompress      string = "gz"
-)
-
-var (
-	// figs is a figtree of fruit for configurable command line arguments that bear fruit
-	figs figtree.Plant
-
-	alwaysWrite = true
-
-	// defaultExclude are the -exc list of extensions that will be skipped automatically
-	defaultExclude = []string{
-		// Compressed archives
-		"7z", "gz", "xz", "zst", "zstd", "bz", "bz2", "bzip2", "zip", "tar", "rar", "lz4", "lzma", "cab", "arj",
-
-		// Encryption, certificates, and sensitive keys
-		"crt", "cert", "cer", "key", "pub", "asc", "pem", "p12", "pfx", "jks", "keystore",
-		"id_rsa", "id_dsa", "id_ed25519", "id_ecdsa", "gpg", "pgp",
-
-		// Binary & executable artifacts
-		"exe", "dll", "so", "dylib", "bin", "out", "o", "obj", "a", "lib", "dSYM",
-		"class", "pyc", "pyo", "__pycache__",
-		"jar", "war", "ear", "apk", "ipa", "dex", "odex",
-		"wasm", "node", "beam", "elc",
-
-		// System and disk images
-		"iso", "img", "dmg", "vhd", "vdi", "vmdk", "qcow2",
-
-		// Database files
-		"db", "sqlite", "sqlite3", "db3", "mdb", "accdb", "sdf", "ldb",
-
-		// Log files
-		"log", "trace", "dump", "crash",
-
-		// Media files - Images
-		"jpg", "jpeg", "png", "gif", "bmp", "tiff", "tif", "webp", "ico", "svg", "heic", "heif", "raw", "cr2", "nef", "dng",
-
-		// Media files - Audio
-		"mp3", "wav", "flac", "aac", "ogg", "wma", "m4a", "opus", "aiff",
-
-		// Media files - Video
-		"mp4", "avi", "mov", "mkv", "webm", "flv", "wmv", "m4v", "3gp", "ogv",
-
-		// Font files
-		"ttf", "otf", "woff", "woff2", "eot", "fon", "pfb", "pfm",
-
-		// Document formats (typically not source code)
-		"pdf", "doc", "docx", "xls", "xlsx", "ppt", "pptx", "odt", "ods", "odp", "rtf",
-
-		// IDE/Editor/Tooling artifacts
-		"suo", "sln", "user", "ncb", "pdb", "ipch", "ilk", "tlog", "idb", "aps", "res",
-		"iml", "idea", "vscode", "project", "classpath", "factorypath", "prefs",
-		"vcxproj", "vcproj", "filters", "xcworkspace", "xcuserstate", "xcscheme", "pbxproj",
-		"DS_Store", "Thumbs.db", "desktop.ini",
-
-		// Package manager and build artifacts
-		"lock", "sum", "resolved", // package-lock.json, go.sum, yarn.lock, etc.
-
-		// Temporary and backup files
-		"tmp", "temp", "swp", "swo", "bak", "backup", "orig", "rej", "patch",
-		"~", "old", "new", "part", "incomplete",
-
-		// Source maps and minified files (usually generated)
-		"map", "min.js", "min.css", "bundle.js", "bundle.css", "chunk.js",
-
-		// Configuration that's typically binary or generated
-		"dat", "data", "cache", "pid", "sock",
-
-		// Version control artifacts (though usually in ignored directories)
-		"pack", "idx", "rev",
-
-		// Other binary formats
-		"pickle", "pkl", "npy", "npz", "mat", "rdata", "rds",
-	}
-
-	// defaultInclude are the -inc list of extensions that will be included in the summary
-	defaultInclude = []string{
-		"go", "ts", "tf", "sh", "py", "js", "Makefile", "mod", "Dockerfile", "dockerignore", "gitignore", "esconfigs", "md",
-	}
-
-	// defaultAvoid are the -avoid list of substrings in file path names to avoid in the summary
-	defaultAvoid = []string{
-		".min.js", ".min.css", ".git/", ".svn/", ".vscode/", ".vs/", ".idea/", "logs/", "secrets/",
-		".venv/", "/site-packages", ".terraform/", "summaries/", "node_modules/", "/tmp", "tmp/", "logs/",
-	}
-)
-
-// newSummaryFilename returns summary.time.Now().UTC().Format(tFormat).md
-var newSummaryFilename = func() string {
-	return fmt.Sprintf("summary.%s.md", time.Now().UTC().Format(tFormat))
-}
-
-// init creates a new figtree with options to use CONFIG_FILE as a way of reading a YAML file while ignoring the env
-func configure() {
-	figs = figtree.With(figtree.Options{
-		Harvest:           9,
-		IgnoreEnvironment: true,
-		ConfigFile:        envVal(eConfigFile, "./config.yaml"),
-	})
-	// properties
-	figs = figs.NewString(kSourceDir, ".", "Absolute path of directory you want to summarize.")
-	figs = figs.NewString(kOutputDir, filepath.Join(".", "summaries"), fmt.Sprintf("Path of the directory to write the %s file to", newSummaryFilename()))
-	figs = figs.NewString(kFilename, newSummaryFilename(), "Output file of summary.md")
-	figs = figs.NewList(kIncludeExt, defaultInclude, "List of extensions to INCLUDE in summary.")
-	figs = figs.NewList(kExcludeExt, defaultExclude, "List of extensions to EXCLUDE in summary.")
-	figs = figs.NewList(kSkipContains, defaultAvoid, "List of path substrings if present to skip over full path.")
-	figs = figs.NewInt(kMaxFiles, 369, "Maximum number of files to process concurrently")
-	figs = figs.NewInt64(kMaxOutputSize, 1_776_369, "Maximum file size of output file")
-	figs = figs.NewBool(kDotFiles, false, "Any path that is considered a dotfile can be included by setting this to true")
-	figs = figs.NewBool(kPrint, envIs(eAlwaysPrint), "Print generated file contents to STDOUT")
-	figs = figs.NewBool(kWrite, envIs(eAlwaysWrite), "Write generated contents to file")
-	figs = figs.NewBool(kJson, envIs(eAlwaysJson), "Enable JSON formatting")
-	figs = figs.NewBool(kCompress, envIs(eAlwaysCompress), "Use gzip compression in output")
-	figs = figs.NewBool(kVersion, false, "Display current version of summarize")
-	figs = figs.NewBool(kDebug, false, "Enable debug mode")
-	// validators
-	figs = figs.WithValidator(kSourceDir, figtree.AssureStringNotEmpty)
-	figs = figs.WithValidator(kOutputDir, figtree.AssureStringNotEmpty)
-	figs = figs.WithValidator(kFilename, figtree.AssureStringNotEmpty)
-	figs = figs.WithValidator(kMaxFiles, figtree.AssureIntInRange(1, 17_369))
-	figs = figs.WithValidator(kMaxOutputSize, figtree.AssureInt64InRange(369, 369_369_369_369))
-	// callbacks
-	figs = figs.WithCallback(kSourceDir, figtree.CallbackAfterVerify, callbackVerifyReadableDirectory)
-	figs = figs.WithCallback(kFilename, figtree.CallbackAfterVerify, callbackVerifyFile)
-}
-
-type result struct {
-	Path     string `yaml:"path" json:"path"`
-	Contents []byte `yaml:"contents" json:"contents"`
-	Size     int64  `yaml:"size" json:"size"`
-}
-
-type final struct {
-	Path     string `yaml:"path" json:"path"`
-	Contents string `yaml:"contents" json:"contents"`
-	Size     int64  `yaml:"size" json:"size"`
-}
 
 func main() {
 	configure()
@@ -471,8 +284,8 @@ func main() {
 		}
 		paths := slices.Clone(thisData.Paths)
 
-		throttler.Acquire() // throttler is used to protect the runtime from excessive use
-		wg.Add(1)           // wg is used to prevent the runtime from exiting early
+		throttler.Acquire()                                                            // throttler is used to protect the runtime from excessive use
+		wg.Add(1)                                                                      // wg is used to prevent the runtime from exiting early
 		go func(innerData *mapData, toUpdate *[]mapData, ext string, paths []string) { // run this extension in a goroutine
 			defer throttler.Release() // when we're done, release the throttler
 			defer wg.Done()           // then tell the sync.WaitGroup that we are done
@@ -582,177 +395,4 @@ func main() {
 			filepath.Join(*figs.String(kOutputDir), *figs.String(kFilename)),
 		)
 	}
-}
-
-var callbackVerifyFile = func(value interface{}) error {
-	return check.File(toString(value), file.Options{Exists: false})
-}
-
-var callbackVerifyReadableDirectory = func(value interface{}) error {
-	return check.Directory(toString(value), directory.Options{Exists: true, MorePermissiveThan: 0444})
-}
-
-var toString = func(value interface{}) string {
-	switch v := value.(type) {
-	case string:
-		return v
-	case *string:
-		return *v
-	default:
-		flesh := figtree.NewFlesh(value)
-		f := fmt.Sprintf("%v", flesh.ToString())
-		return f
-	}
-}
-
-var capture = func(msg string, d ...error) {
-	if len(d) == 0 || (len(d) == 1 && d[0] == nil) {
-		return
-	}
-	terminate(os.Stderr, "[EXCUSE ME, BUT] %s\n\ncaptured error: %v\n", msg, d)
-}
-
-type m struct {
-	Message string `json:"message"`
-}
-
-var terminate = func(d io.Writer, i string, e ...interface{}) {
-	for _, f := range os.Args {
-		if strings.HasPrefix(f, "-json") {
-			mm := m{Message: fmt.Sprintf(i, e...)}
-			jb, err := json.MarshalIndent(mm, "", "  ")
-			if err != nil {
-				_, _ = fmt.Fprintf(os.Stderr, "Error serializing json: %v\n", err)
-				_, _ = fmt.Fprintf(d, i, e...)
-			} else {
-				fmt.Println(string(jb))
-			}
-			os.Exit(1)
-		}
-	}
-	_, _ = fmt.Fprintf(d, i, e...)
-	os.Exit(1)
-}
-
-func simplify(t []string) []string {
-	seen := make(map[string]bool)
-	for _, v := range t {
-		seen[v] = true
-	}
-	results := make([]string, len(t))
-	for i, v := range t {
-		if seen[v] {
-			results[i] = v
-		}
-	}
-	return results
-}
-func addFromEnv(e string, l *[]string) {
-	v, ok := os.LookupEnv(e)
-	if ok {
-		flesh := figtree.NewFlesh(v)
-		maybeAdd := flesh.ToList()
-		for _, entry := range maybeAdd {
-			*l = append(*l, entry)
-		}
-	}
-	*l = simplify(*l)
-}
-
-type seenStrings struct {
-	mu sync.RWMutex
-	m  map[string]bool
-}
-
-func (s *seenStrings) Add(entry string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.m[entry] = true
-}
-func (s *seenStrings) Remove(entry string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	delete(s.m, entry)
-}
-
-func (s *seenStrings) Len() int {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return len(s.m)
-}
-
-func (s *seenStrings) String() string {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return fmt.Sprint(s.m)
-}
-
-func (s *seenStrings) True(entry string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.m[entry] = true
-}
-
-func (s *seenStrings) False(entry string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	delete(s.m, entry)
-}
-
-func (s *seenStrings) Exists(entry string) bool {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.m[entry]
-}
-
-func envVal(name, fallback string) string {
-	v, ok := os.LookupEnv(name)
-	if !ok {
-		return fallback
-	}
-	return v
-}
-
-func envIs(name string) bool {
-	v, ok := os.LookupEnv(name)
-	if !ok {
-		return false
-	}
-	vb, err := strconv.ParseBool(v)
-	if err != nil {
-		return false
-	}
-	return vb
-}
-
-// compress compresses a string using gzip and returns the compressed bytes
-func compress(s []byte) ([]byte, error) {
-	var buf bytes.Buffer
-	gzWriter := gzip.NewWriter(&buf)
-	_, err := gzWriter.Write(s)
-	if err != nil {
-		return nil, fmt.Errorf("failed to write to gzip writer: %w", err)
-	}
-	err = gzWriter.Close()
-	if err != nil {
-		return nil, fmt.Errorf("failed to close gzip writer: %w", err)
-	}
-	return buf.Bytes(), nil
-}
-
-// decompress decompresses gzip compressed bytes back to a string
-func decompress(compressed []byte) (string, error) {
-	buf := bytes.NewReader(compressed)
-	gzReader, err := gzip.NewReader(buf)
-	if err != nil {
-		return "", fmt.Errorf("failed to create gzip reader: %w", err)
-	}
-	defer func() {
-		_ = gzReader.Close()
-	}()
-	decompressed, err := io.ReadAll(gzReader)
-	if err != nil {
-		return "", fmt.Errorf("failed to read from gzip reader: %w", err)
-	}
-	return string(decompressed), nil
 }
